[Feat] xml 파싱완료

+++ Easy_Route_Assistant/src/main/java/com/founder/easy_route_assistant/Entity/BusStationEntity.java
@@ -14,13 +14,13 @@ import lombok.*;
 public class BusStationEntity {
     @Id
     @Column(name = "bus_route_id")
-    Long busRouteId;
+    Integer busRouteId;
     @Column(name = "bus_name")
     String busName;
     @Column
-    Long ord;
+    Integer ord;
     @Column(name = "st_id")
-    Long stId;
+    Integer stId;
     @Column(name = "st_name")
     String stName;
